+++ workshop/Contents/mods/AuxiliasCrossbow/42.20/media/registries.lua
@@ -1,2 +1,5 @@
 local boltKey = ItemKey.new("AuxiliasCrossbowBolt", ItemType.NORMAL)
 AmmoType.register("auxiliascrossbow:bolt", boltKey:toString())
+
+local stoneBoltKey = ItemKey.new("AuxiliasStoneCrossbowBolt", ItemType.NORMAL)
+AmmoType.register("auxiliascrossbow:stonebolt", stoneBoltKey:toString())